# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.CreateAdminRequest:
    properties:
      name:
        example: Alisher Karimov
        maxLength: 150
        minLength: 2
        type: string
      password:
        example: secret123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901234567"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: admin
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - password
    - phone
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse:
    properties:
      error:
        example: validation failed
        type: string
      message:
        example: Xatolik yuz berdi
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.LoginRequest:
    properties:
      password:
        example: secret123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse:
    properties:
      data: {}
      message:
        example: Muvaffaqiyatli
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.UpdateAdminRequest:
    properties:
      name:
        example: Alisher Karimov
        maxLength: 150
        minLength: 2
        type: string
      password:
        example: newsecret123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901234567"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: admin
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - phone
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.UpdateProfileRequest:
    properties:
      name:
        example: Alisher Karimov
        maxLength: 150
        minLength: 2
        type: string
      password:
        example: newsecret123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901234567"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: admin
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - phone
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Modular monolit backend. Admin, Company, Survey va Settings modullari.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Foydalanuvchilar bazasi API
  version: "1.0"
paths:
  /admins:
    get:
      description: Barcha adminlarni qaytaradi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adminlar ro'yxati
      tags:
      - Admins
    post:
      consumes:
      - application/json
      description: 'Admin CRUD: yangi admin qo''shadi'
      parameters:
      - description: Admin ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Yangi admin yaratish
      tags:
      - Admins
  /admins/{id}:
    delete:
      description: Adminni tizimdan o'chiradi
      parameters:
      - description: Admin UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adminni o'chirish
      tags:
      - Admins
    get:
      description: Bitta admin ma'lumotlarini qaytaradi
      parameters:
      - description: Admin UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adminni ID bo'yicha olish
      tags:
      - Admins
    put:
      consumes:
      - application/json
      description: Admin ma'lumotlarini o'zgartiradi. Password bo'sh qoldirilsa, eski
        parol saqlanadi.
      parameters:
      - description: Admin UUID
        in: path
        name: id
        required: true
        type: string
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adminni yangilash
      tags:
      - Admins
  /auth/login:
    post:
      consumes:
      - application/json
      description: Username va password orqali tizimga kirish. JWT token qaytaradi.
      parameters:
      - description: Login ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      summary: Admin login
      tags:
      - Auth
  /auth/profile:
    get:
      description: Token egasining profil ma'lumotlarini qaytaradi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Joriy admin profili
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Token egasi o'z profilini o'zgartiradi. Password bo'sh qoldirilsa,
        eski parol saqlanadi.
      parameters:
      - description: Profil ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_admin_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Joriy admin profilini yangilash
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"